==> qm9-emissions/0_data_selection.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import typing
from rdkit import Chem
from rdkit.Chem import Draw
from rdkit.Chem import Descriptors
from typing import Union, List, Tuple
from rdkit.Chem import AllChem
from rdkit.Chem import Descriptors3D
from rdkit.Chem import GraphDescriptors
from tqdm import tqdm

### Helper functions

def count_atoms(smils: Chem.Mol, atomic_value: int) -> int:
    count = 0
    smilsChromoChem = smils
    for atom in smilsChromoChem.GetAtoms():
        if atom.GetAtomicNum() == atomic_value:
            count += 1
    return count

def find_pattern_indexes(molecule: Chem.Mol, patterns_dataframe: pd.DataFrame) -> List[int]:
    ret = []
    for index, row in patterns_dataframe.iterrows():
        if len(list(molecule.GetSubstructMatch(row["RDKit_Mol_Class"])))>0:
            ret.append(index)
    return ret

## Read databases

chromophore_database = pd.read_csv("DB for chromophore_Sci_Data_rev02.csv")
chromophore_database.drop("Tag", inplace=True, axis=1)
chromophore_database = chromophore_database.iloc[:, 0:4]
chromophore_database.drop(columns="Absorption max (nm)", inplace=True)
only_solids = chromophore_database[chromophore_database["Chromophore"]==chromophore_database["Solvent"]].copy()
only_solids["RDKit_Mol_Class"] = only_solids["Chromophore"].apply(lambda x: Chem.MolFromSmiles(x))
qm9_database = pd.read_csv("qm9.csv")

### Find patterns from QM9 database
Needs time.

qm9_database["RDKit_Mol_Class"] = qm9_database["smiles"].apply(lambda x: Chem.MolFromSmiles(x))
only_solids["qm9_pattern_indexes"] = only_solids["RDKit_Mol_Class"].apply(find_pattern_indexes, patterns_dataframe = qm9_database)

### Count number of each atoms of N, O, S, F, Cl, Br, I, Se, Te, Si, P, B, Sn, Ge

for atom, value in tqdm(zip(
    [x.strip() for x in "N, O, S, F, Cl, Br, I, Se, Te, Si, P, B, Sn, Ge".split(",")],
    [7, 8, 16, 9, 17, 35, 53, 34, 52, 14, 15, 5, 50, 32]
)):
    only_solids[f"{atom}_count"] = only_solids["RDKit_Mol_Class"].apply(count_atoms, atomic_value=value)

only_solids.head()

only_solids.info()

### Create subset of chromophores consisting of C, O, N, F atoms

only_solids_conf = only_solids.copy()
for atom in tqdm([x.strip() for x in "S, Cl, Br, I, Se, Te, Si, P, B, Sn, Ge".split(",")]):
    only_solids_conf.drop(only_solids_conf[only_solids_conf[f"{atom}_count"]>0].index, inplace=True)
    only_solids_conf.drop(columns=f"{atom}_count", inplace=True)

only_solids_conf.head()

only_solids_conf.info()

## Save databases

only_solids.drop(columns=["RDKit_Mol_Class", "Solvent"]).to_json("only_solids.json")
only_solids_conf.drop(columns=["RDKit_Mol_Class", "Solvent"]).to_json("only_solids_conf.json")
